requirements-dev.txt
-e .
coverage==4.4.1
isort==4.2.5
py==1.4.33
pytest-django==3.1.2
pytest==3.0.7
selenium==3.4.2
